Split as client component
